moved tests to one file
