+++ components/CarouselItem.js
@@ -28,7 +28,6 @@ import Link from 'next/link';
 //     }
 // }
 export default function carouselItem({ restoSet }) {
-    console.log("CAR ITEM");
     return (  
         <Row justify="space-around">
             {restoSet.map ((card, index) => {
